[trace-host] move trace related logic to separate module
- This helps with keeping code separate and easily gating it out
